Bugs item #877873 Flash link extractor causes OutOfMemory exceptions.
Rewrote ExtractorSWF so that it does not use excessive amount of memory while parsing SWF files.
The old version of swf extractor was using writer�s tags implementation that was then used by the reader to parse URIs. The writer it self keeps local buffer of an entire swf file that is at the end flushed to an output stream (in this case null.) So while parsing large swf files, extractor was causing OutOfMemomry exceptions especially on the systems with limited amount of memory available.
The new version simple overwrites concrete implementation of tags found within the API that is than passed to the reader. This way minimum amount of memory is used while parsing one tag at the time.