Added functionality that addresses RFE #935215 day/night configurations.
Added feature for adding refinements to a setting (global or per).
A refinement has criteria connected to it that a URI must pass for
the refinements to override the settings. The criteria types are:
Time of day, uri matches regular expression or uri has a spesific port number.

UI-support needs to be added.