Fix for [ 1060589 ] Can't open logs of old jobs post-restart in UI Patch submitted by Kris.  Here is his email.
A couple of things.

1. Michael's change causes some problems. By rewriting fileName (the
variable) to also include the path, other things get broken. Namely the
'having the current log selected (black text, no underscore)' and the
display of the current logs name (if we want to display the path, it should
be done somewhere at the bottom of the page, not in the 'log display
header'.)

2. To fix all of this how about adding a 'getLogsPath()' to CrawlJob. The
code in getSettingsDirectory() could be replicated (it's just reading from
the CrawlOrder) minus the 'create dir if it doesn't exist already'.

That would have the additional benefit of removing a bit more 'code' from
the JSP pages (usually a good thing).
So the above section in the logs.jsp file would just be:

So whenever you need the absolute path of a log you use
theJob.getLogPath(fileName);

See patches below. (As far as I can tell it works under all the
circumstances Michael cited)

* src/java/org/archive/crawler/admin/CrawlJob.java
    (getLogPath): Added.
* src/webapps/admin/logs.jsp
    Use new getLogPath method.
* src/articles/releasenotes.xml
    Removed logs not working on old crawls as known problem.