From Linda Lee at bdb:
"Attached is a jar which we believe will address the "thread was notified, but
is not owner" exception you've seen. For the record, it's CVS tag is
archive_org_12115_11552_2nd.  The essence of the change is that we've convinced
ourselves that the sanity check is spurious and that the latch code itself is
sound."

* .classpath
* project.properties
    Update for new je jar.
* lib/je-1.7.1-12115_11552-2.jar
    Added. Replaces...
* lib/je-1.7.1-12115_11552.jar
    Removed.