Implement '[ 1388295 ] [contrib] Throttling on a per-document basis' Contribution from Karl Wright (karl at metadata dot com).  Below is commit message from Karl.
"Addition of a "fetch-bandwidth" HTTP fetcher expert setting that allows one to
control the maximum burst fetch rate for any single document fetch.  This is
particularly helpful for controlling how much of a server's actual T1
bandwidth Heritrix uses, even when it is fetching large, multimegabyte
documents.  We added this capability as a result of receiving mail from people
we were crawling who ran a library, and who were negatively impacted because of
our crawl.

"The default value is 0, which means no limit.  We set it to a fast cable rate
(384KB/second).  We have found that even this is a bit much for some servers
out there, who prefer a burst rate more like 64KB/second."

Applied by St.Ack.  Formatting changes and changed default form 384 to 0 i.e.
no limit.  Made it an expert setting and overrideable.  Added javadoc for new
parameters on methods.

* src/java/org/archive/crawler/fetcher/FetchHTTP.java
	(ATTR_FETCH_BANDWIDTH_MAX, DEFAULT_FETCH_BANDWIDTH_MAX,
	    getMaxFetchRate): Added.
* src/java/org/archive/io/RecordingInputStream.java
	(readFullyOrUntil): Added maxBytesPerMs param.
* src/java/org/archive/io/RecordingInputStreamTest.java
	Changed API.