Switched to jetty5. Apparently there's a bug in jetty6 which was preventing my servlets from receiving POST data. The tester creates a task and the monkey successfully retrieves it.
moved configuration to external config file. It is created with default values in ~/.archive_monkey when the controller is run for the first time

created ant build file.

run ant dist to create a distribution then use start_controller.sh in dist to run the controller

servlets are still missing some of the methods but it's trivial to add (i'm doing it)