Refactored FetchHTTP to use new Key-based settings and also to isolate it from the crawler.
About a third of the code that once lived in FetchHTTP is now commented out. Cookie management is now delegated to a configurable CookieStorage module. This prevents FetchHTTP from needing its own custom serialization code. Credential handling was lifted into ProcessorURI abstractions.

Also there is a somewhat major bug fix here. The previous implementations of HeritrixProtocolSocketFactory and HeritrixSSLProtocolSocketFactory wouldn't work properly if there was more than one crawler in a single JVM. The second registered ProtocolFactory would replace the first, so the first crawler would end up using the second crawler's ServerCache.

So I changed the ProtocolFactory stuff to use the current thread to resolve host names; the current thread must implement HostResolver for that to work. So, ToeThread will now be a HostResolver.

That change also prevents FetchHTTP from needing a ServerCache.

We still need a BDB and flat file implementation of CookieStorage.

* /src/java/org/archive/crawler/fetcher/FetchHTTP.java
Refactored; uses Key fields instead of ATTR_ and DEFAULT_ constants. Uses ProcessorURI instead of CrawlURI. Delegates cookie storage to a CookieStorage module.

* /src/java/org/archive/crawler/fetcher/CookieStorage.java
New module for saving and loading cookies.

* /src/java/org/archive/crawler/fetcher/SimpleCookieStorage.java
Simple implementation of CookieStorage that just keeps cookies in memory. Can be serialized, though, so probably useful for small crawls.

* /src/java/org/archive/crawler/fetcher/HostResolver.java
New marker interface for threads that can resolve hosts.

* /src/java/org/archive/crawler/fetcher/HeritrixProtocolSocketFactory
* /src/java/org/archive/crawler/fetcher/HeritrixSSLProtocolSocketFactory
Now checks the current thread to see if it's a HostResolver, instead of relying on a local ServerCache.