Altered KeyManager so that instead of having one flag that determines whether or not a setting is "overrideable", there are several flags that determine the setting's readability and writeability:
1. Normal: The setting is readable and writeable in any context.
2. Global: The setting is readable and writeable only in the global
context.
3. Immutable: The setting is only readable, and only in the global
context.
4. Dependency: The setting is only readable, and only in the global
context, and the setting value must be fully constructed before the
dependent object can be created.

Rather than coming up with methods like "makeGlobal" and
"makeExpertDependency", or to provide fewer methods that take seven
parameters to cover all the cases, I created annotations to indicate a
static Key field's scope.  The annotations are:

@Global
@Immutable
@Dependency

They are used when a key is declared, as in:

   @Global
   final public static Key<String> DIR = Key.make(".");

Additionally, there is an @Expert tag that indicates whether or not a
Key is expert.

The use of annotations has the additional benefit that these annotations
show up in the generated JavaDoc for Key fields.

So the KeyManager.addKeys method was overhauled to look for these
annotations and set flags within the Key objects appropriately.

I added isGlobal, isImmutable and isDependency accessors to Key.

I removed methods from Key and KeyMaker that looked like "makeExpert", 
"makeExpertFinal" and so on.

Additionally, the org.archive.state package now enforces that only
certain types are allowed.  These types are listed in the new KeyTypes
class. Previously I thought that upstream frameworks (like
org.archive.settings) could specify which types they do or do not
support, but it became necessary to specify it here so that dependencies
could be tracked correctly.

* /src/java/org/archive/state/KeyManager.java
(addKeys):  Now checks for annotations and sets flags accordingly.

* /src/java/org/archive/state/Key.java
(setMetadata):  Now applies flags based on annotations present in the
java.lang.reflect.Field that represents the Key.
(isDependency)
(isGlobal)
(isImmutable):  New accessors for determining a key's scope.

* /src/java/org/archive/state/KeyTypes.java
New class for specifying which types are allowed for simple keys.

* /src/java/org/archive/state/Dependency.java
* /src/java/org/archive/state/Expert.java
* /src/java/org/archive/state/Global.java
* /src/java/org/archive/state/Immutable.java
New annotations for marking up static Key fields.