Added support for "offline" sheet managers, which hold settings and information for objects without actually constructing those objects. This is needed by the org.archive.crawler.admin stuff, to be able to construct a new crawler based on some offline profile.
Additionally, moved the merging of data structures (lists and maps) into
org.archive.settings.  I thought I could get away with delaying this
merging until the very last possible moment (when you have a URI in SURT
form and you've queried the SheetManager for a sheet associated with
every piece of the SURT), but that turned out not to be workable, as the
admin packages need to be able to produce detailed information on sheets
even if those sheets are in "offline" mode mentioned above.

Finally, introduced the notion of the "unspecified" sheet.  Previously,
the "default" sheet contained both information actually specified in a
file for the default sheet, and also, if that information were not
available, the default values as determined by Key.getDefaultValue().
These are now two separate concepts; if an operator does not explicitly
define a setting in the default sheet, then the value for that setting
is said to have come from the "unspecified" sheet.  This was an easy
change that helped during unit testing, and I imagine it will help
operators on live crawls as well.

I modified the unit tests to examine dependencies, data structure
merging, both online and offline sheets.  The tests could probably be
more comprehensive but I'm pretty sure things work now.

Not going into details on every change since more or less everything was touched in these packages:

org.archive.settings
org.archive.settings.path