Fixed memory leak in SingleSheet.  Made it so that the presence of settings for a module will not prevent that module from being garbage collection.  And when a module is garbage collected, its settings are automatically removed from any sheets that configured that module.
* /src/java/org/archive/settings/ModuleReference.java
New subclass of PhantomReference that indicates the sheet and identity
of a module.  When the module becomes phantom reachable, a ModuleReference
is used to remove its settings from the single sheet.

* /src/java/org/archive/settings/SingleSheetCleanUpThread.java
New daemon thread that removes ModuleReferences from a ReferenceQueue
and performs cleanup.

* /src/java/org/archive/settings/SingleSheet.java
(settings): Is now a ConcurrentMap.  Now uses the integer identity hash
code of a module instead of the module itself as the key.
(check)
(checkOffline): Now uses integer identity hash code for lookup in
settings map.
(set): Uses integer identity hash code for lookup; uses ConcurrentMap
features to ensure thread safety during concurrent writes; registers
modules with SingleSheetCleanupThread when they are first added to the
settings map.