* .classpath * commons/pom.xml Bump sleepycat je from 3.2.13 to 3.2.23.
* heritrix/src/main/java/org/archive/crawler/util/BdbUriUniqFilter.java
(<init>): Set allowCreate to true instead of false.
(close): Move test for null environment before environment is manually
set to null.  The bad test was causing database/env to not be close()'d
during unit tests.

* heritrix/src/main/java/org/archive/crawler/util/BdbUriUniqFilterTest.java
(testForget): Renamed to testForgetOnEmpty, which is a little more accurate.
Added TODO for a real testForget.