Made the sheet editor webui page a bit smarter about what can and cannot be changed in a sheet.
* commons/src/main/java/org/archive/settings/jmx/JMXSheetManager.java
* commons/src/main/java/org/archive/settings/jmx/JMXSheetManagerImpl.java
(isOnline): New method for determining whether a sheet manager is online.

* webui/src/main/java/org/archive/crawler/webui/Setting.java
(key)
(getKey)
(setKey)
(isKeySet): New methods for caching a setting's Key, if it has one.

* webui/src/main/java/org/archive/crawler/webui/Settings.java
(<init>): Takes new parameter that determines if the sheet manager that 
produced the settings is online or not.
(getKey): New method for determining a setting's key.
(getDescription): Now uses getKey.
(isEnabled): New method for determining whether a setting's form field
should be enabled. Takes into account whether the sheet manager is
online or offline, and whether or not the setting/sheet is global.
(printSelectOpenTag)
(printTextField): Now uses isEnabled instead of just checking the sheet name.
(canOverride): New method for determining whether a setting can be overriden.

* webui/src/main/java/org/archive/crawler/webui/Sheets.java
(getSettings): Now sets the online flag in the Settings object.

* webui/src/test/java/org/archive/crawler/webui/WebUIJUnit.java
(testWebui): Changed recorder-in-bufffer-bytes override to
HTTP:timeout-seconds override, since recorder-in-buffer-bytes can't be
overridden.

* webui/src/main/webapp/sheets/page_sheet_editor.jsp
Uses the new canOverride method to determine whether the add/remove
links should be displayed.

* webui/src/main/webapp/crawler_area/page_crawler.jsp
Now includes a link to the sheets for active jobs.