Moved production of final tarball from the heritrix subproject to the parent project, so that it can include the webui .war file.
* pom.xml
Now includes a maven-assembly-plugin for making the final tarball.

* src/main/assembly/dist.xml
New assembly descriptor for the final tarball; based on the old one in 
heritrix/src/main/assembly.

* heritrix/src/site
* heritrix/src/main/bin
* heritrix/src/main/conf
* heritrix/README.txt
* heritrix/LICENSE.txt
Moved these from heritrix/ to ./, so the parent assembly-plugin can find them.

* heritrix/src/test/java/org/archive/settings/file/ValidatorTest.java
Changed where the test looks for the default profile, since it moved. 
Also, this test now copies the default profile to /tmp instead of
opening it in-place; this prevents a state directory from appearing in
the src tree.

* heritrix/src/main/assembly/dist.xml
No longer produces the final tarball; it DOES still produce a list of
all the dependency jars, since it's apparently impossible to do that in
the parent pom.