* heritrix/src/main/java/org/archive/crawler/WebUI.java New utility for running the web ui in a Jetty.  Mostly I wanted to isolate this code from Heritrix.main().
* heritrix/src/main/java/org/archive/crawler/WebUIConfig.java
The paramters needed to start a web ui.

* heritrix/src/main/java/org/archive/crawler/Heritrix.java
Now parses four new command line arguments: -r means to run the 
web ui, -w specifies a path to the webui WAR file, -b specifies
a list of hostnames for the webui to bind to, and -l specifies
the port for the webui to listen on.  Those last three have no
effect unless -r is specified, and they have sane defaults if
-r is specified but the other three are not.