Changed how opening/closing of JMXSheetManagers is handled.
* commons/src/main/java/org/archive/settings/jmx/JMXSheetManagerImpl.java
Now has awareness of own ObjectName

* engine/src/main/java/org/archive/crawler/framework/CrawlJobManager.java
openProfile method renamed to getProfile and now returns the ObjectName of the JMXSheetManager for the profile

* engine/src/main/java/org/archive/crawler/framework/CrawlJobManagerImpl.java
All public methods are now synchronized.
Changed opening and closing of JMXSheetManagers so that CJM ensures that multiple instances are never created for the same profile.

* webui/src/main/java/org/archive/crawler/webui/Sheets.java
Changed (simplified) how JMXSheetManagers are accessed to conform with changes made on the server side.