Fix for HER-1349: Test URL with empty or random input
StoredSortedSet was returning the wrong thing for headSet(str + '\0').
The returned set was including the next element AFTER str.  I think 
this might be because they use \0 to delineate records in the DB.

* commons/src/main/java/org/archive/settings/file/PrefixFinder.java
(find): Use \u0001 instead of \0 as the extra character.  Include
comment about why we're doing this strange thing.

* commons/src/test/java/org/archive/settings/file/PrefixFinderTest.java
(testStrangeCase): New test case for making sure this bug is fixed.